Refactored and added Proptypes to Navbar drop down menu
