+++ client/src/components/Mmpnavigation/Dropdown/index.jsx
@@ -1,0 +1,42 @@
+import React, { useState, cloneElement } from "react";
+import PropTypes from "prop-types";
+
+export default function Dropdown({ trigger, menu }) {
+  const [open, setOpen] = useState(false);
+
+  const handleOpen = () => {
+    setOpen(!open);
+  };
+
+  return (
+    <div className="dropdown">
+      {cloneElement(trigger, {
+        onClick: handleOpen,
+      })}
+      {open ? (
+        <ul className="menu">
+          {menu.map((menuItem, index) => (
+            <li key={index} className="menu-item">
+              {cloneElement(menuItem, {
+                onClick: () => {
+                  menuItem.props.onClick();
+                  setOpen(false);
+                },
+              })}
+            </li>
+          ))}
+        </ul>
+      ) : null}
+    </div>
+  );
+}
+
+Dropdown.propTypes = {
+  trigger: PropTypes.node.isRequired,
+  menu: PropTypes.arrayOf(
+    PropTypes.shape({
+      menuItem: PropTypes.element.isRequired,
+      index: PropTypes.string.isRequired,
+    })
+  ).isRequired,
+};
