+++ client/src/components/Mmpnavigation/index.jsx
@@ -1,12 +1,30 @@
 import React from "react";
 import { Link } from "react-router-dom";
 import styles from "./Navigation.module.css";
-import Hamburger from "./Hamburger";
+import Dropdown from "./Dropdown";
 
 export default function Navigation() {
+  const handleMenuOne = () => {
+    console.log("clickedOne");
+  };
+
+  const handleMenuTwo = () => {
+    console.log("clicked two");
+  };
+
   return (
     <ul className={styles.container}>
-      <Hamburger />
+      <Dropdown
+        trigger={<button type="button">Dropdown</button>}
+        menu={[
+          <button type="button" onClick={handleMenuOne}>
+            Menu 1
+          </button>,
+          <button type="button" onClick={handleMenuTwo}>
+            Menu 2
+          </button>,
+        ]}
+      />
       <Link to="/">eRenaissance</Link>
       <Link to="/profile">Profile Icon</Link>
       <Link to="/basket">Basket Icon</Link>
